Change List to a circular linked list
Added a Node class to represent a data node in a list. Replaced the
internal data storage with a linked list structure. The list is now
circular and singly linked.

Warning: the tests in this commit will not work because the methods
are not adapted to the new type of list.