Update length method for circular linked list
Update length method to implement traversal logic to count
nodes in a circular linked list. Tests for methods that have not yet
been refactored are temporarily commented out. Tests for the length method
have also been temporarily disabled, as this method requires the delete method
to work correctly, which will be provided in the next commit.